+++ Source/API/Table/Content/NSAttributedString+PDFTableContent.swift
@@ -1,0 +1,21 @@
+//
+//  NSAttributedString+PDFTableContent.swift
+//  TPPDF
+//
+//  Created by Philip Niedertscheider on 20.12.19.
+//
+
+import Foundation
+
+/**
+ TODO: Documentation
+ */
+public extension NSAttributedString {
+
+    /**
+     TODO: Documentation
+     */
+    func toPDFTableContent() -> PDFTableContent {
+        return PDFTableContent(type: .attributedString, content: self)
+    }
+}
