+++ Source/API/Table/Content/PDFTableContent.swift
@@ -143,42 +143,3 @@ public class PDFTableContent: PDFJSONSerializable {
         return type == .image ? content as? UIImage : nil
     }
 }
-
-/**
- TODO: Documentation
- */
-public extension String {
-
-    /**
-     TODO: Documentation
-     */
-    func toPDFTableContent() -> PDFTableContent {
-        return PDFTableContent(type: .string, content: self)
-    }
-}
-
-/**
- TODO: Documentation
- */
-public extension NSAttributedString {
-
-    /**
-     TODO: Documentation
-     */
-    func toPDFTableContent() -> PDFTableContent {
-        return PDFTableContent(type: .attributedString, content: self)
-    }
-}
-
-/**
- TODO: Documentation
- */
-public extension UIImage {
-
-    /**
-     TODO: Documentation
-     */
-    func toPDFTableContent() -> PDFTableContent {
-        return PDFTableContent(type: .image, content: self)
-    }
-}
