+++ Source/API/Table/Content/UIImage+PDFTableContent.swift
@@ -1,0 +1,21 @@
+//
+//  UIImage+PDFTableContent.swift
+//  TPPDF
+//
+//  Created by Philip Niedertscheider on 20.12.19.
+//
+
+import Foundation
+
+/**
+ TODO: Documentation
+ */
+public extension UIImage {
+
+    /**
+     TODO: Documentation
+     */
+    func toPDFTableContent() -> PDFTableContent {
+        return PDFTableContent(type: .image, content: self)
+    }
+}
