+++ Source/API/Table/Style/PDFTableStyle+Defaults.swift
